requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
Flask==3.0.3
gunicorn==22.0.0

# ── Database ──────────────────────────────────────────────────────────────────
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7

# ── Auth ──────────────────────────────────────────────────────────────────────
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1

# ── Rate Limiting ─────────────────────────────────────────────────────────────
Flask-Limiter==3.7.0

# ── Background Jobs ───────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.7

# ── AI ────────────────────────────────────────────────────────────────────────
anthropic==0.30.1

# ── Scraping ──────────────────────────────────────────────────────────────────
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2

# ── Image Processing ─────────────────────────────────────────────────────────
Pillow==10.4.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
